--- lib/XDColors.dart
@@ -1,3 +1,0 @@
-import 'package:flutter/material.dart';
-
-class XDColors {}
